refactor: from CJS to ES6 , done
